BOTTOM NAVIGATION IN ANDROID
Android Bottom Navigation stays at the bottom of the screen providing navigation between top-level views in the app. This is introduced in design support library with backward compatibility. Bottom Navigation should be used when the app has three to five top-level navigations

This repo simplified the process of designing a bottom navigation which hides itself when user scrolls and automatic pop up on scroll stop